feat: build favourites ui
