+++ lib/resources/localization/l10n/intl_en.arb
@@ -64,7 +64,10 @@
     "meat": "Meat",
     "chicken": "Chicken",
     "orange": "Orange",
+    "bestSeller":"Best Seller",
+    "notification_comment": "--------------------------- Notifications -------------------------------",
     "notifications": "Notifications",
-    "bestSeller":"Best Seller"
+    "favourites_comment": "--------------------------- favourites -------------------------------",
+    "favourites": "Favourites"
 
 }
